# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Eden Life",
        "description": "Welcome to the Garden of Eden. it's new day, another opportunity to be a step closer to 'Paradise'!",
        "contact": {
            "email": "olusolagloryolamide@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/customers": {
            "get": {
                "tags": [
                    "customers"
                ],
                "summary": "Display all Customers and their respective gardener",
                "description": "Get all customers and their respective gardener",
                "operationId": "customers",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "customers"
                ],
                "summary": "Register Customers",
                "description": "Add a new customer to the system",
                "operationId": "customerRegister",
                "requestBody": {
                    "description": "Pass customers details (Location-Area and Country must be code-name returned from each of their endpoint)",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "email",
                                    "location_area",
                                    "country_of_domicile"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "Firstname Lastname"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "customer@mail.com"
                                    },
                                    "location_area": {
                                        "type": "string",
                                        "example": "LA"
                                    },
                                    "country_of_domicile": {
                                        "type": "string",
                                        "example": "NG"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "406": {
                        "description": "Not acceptable"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "errors": {
                                            "properties": {
                                                "full_name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": [
                                                            "The full name field is required."
                                                        ]
                                                    },
                                                    "collectionFormat": "multi"
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": [
                                                            "The email field is required."
                                                        ]
                                                    },
                                                    "collectionFormat": "multi"
                                                },
                                                "location_area": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": [
                                                            "The location area field is required."
                                                        ]
                                                    },
                                                    "collectionFormat": "multi"
                                                },
                                                "country_of_domicile": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": [
                                                            "The country of domicile field is required."
                                                        ]
                                                    },
                                                    "collectionFormat": "multi"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/customers/{customerId}": {
            "get": {
                "tags": [
                    "customers"
                ],
                "summary": "Customer by id and their gardener",
                "description": "Get customer by id and their gardener",
                "operationId": "customersById",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "description": "ID of customer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Customer with given ID not found"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}